requirements.txt
altgraph==0.17.4
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
